permissions/admin_permission.py
from rest_framework import permissions

class CustomAdminPermission(permissions.BasePermission):
    def has_permission(self, request, view):
        return request.user.is_authenticated and request.user.is_superuser